pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "finrulebench"
version = "0.1.0"
description = "Sandboxed financial reasoning benchmark for AI models"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.8",
  "typer>=0.12",
  "pyyaml>=6.0",
  "pandas>=2.2",
  "httpx>=0.27",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "mypy>=1.10",
]
openai = [
  "openai>=1.0",
]

[project.scripts]
finrulebench = "finrulebench.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false
